Error related constants & util move to own files in new Error dir in core
