+++ src/components/Card/triggerBinLookUp.ts
@@ -1,6 +1,7 @@
 import fetchJSONData from '../../utils/fetch-json-data';
 import { CbObjOnError } from '../internal/SecuredFields/lib/types';
-import { ERROR_MSG_UNSUPPORTED_CARD_ENTERED, getError } from '../internal/SecuredFields/lib/configuration/constants';
+import { getError } from '../../core/Errors/utils';
+import { ERROR_MSG_UNSUPPORTED_CARD_ENTERED } from '../../core/Errors/constants';
 
 export default function triggerBinLookUp(callbackObj) {
     // Allow way for merchant to disallow binLookup by specifically setting the prop to false
