+++ src/core/Errors/utils.ts
@@ -1,0 +1,19 @@
+import { ERROR_CODES } from './constants';
+
+/**
+ * Access items stored in the ERROR_CODES object by either sending in the key - in which case you get the value
+ * or by sending in the value - in which case you get the key
+ * @param keyOrValue - key (or value) by which to retrieve the corresponding value (or key)
+ */
+export const getError = (keyOrValue: string): string => {
+    // Retrieve value
+    let rtnVal = ERROR_CODES[keyOrValue];
+    if (rtnVal) return rtnVal;
+
+    // Retrieve key
+    rtnVal = Object.keys(ERROR_CODES).find(key => ERROR_CODES[key] === keyOrValue);
+    if (rtnVal) return rtnVal;
+
+    // Neither exist
+    return null;
+};
